requirements.txt
configparser==3.5.0
Django==1.11.5
django-configurations==2.0
django-extensions==1.9.1
django-filter==1.0.4
djangorestframework==3.6.4
Markdown==2.6.9
mysqlclient==1.3.12
Pygments==2.2.0
pytz==2017.2
six==1.11.0
